# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bihar Teacher Mutual Transfer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface OtpSentResponse {
  message: string;
  /** OTP expiry in seconds */
  expiresIn: number;
}

export interface MobileOtpRequest {
  /** 10-digit mobile number */
  phone: string;
}

export interface MobileOtpVerify {
  phone: string;
  otp: string;
}

export interface EmailOtpRequest {
  /** Teacher email address */
  email: string;
}

export interface EmailOtpVerify {
  /** Teacher email address */
  email: string;
  otp: string;
}

export interface Teacher {
  id: number;
  name: string;
  employeeId: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** e.g. Primary Teacher, Middle Teacher, High School Teacher */
  designation: string;
  /** @nullable */
  subject?: string | null;
  district: string;
  block: string;
  school: string;
  /** @nullable */
  schoolCode?: string | null;
  /** @nullable */
  serviceYears?: number | null;
  isProfileComplete: boolean;
  createdAt: string;
}

export interface AuthResponse {
  teacher: Teacher;
  isNewUser: boolean;
}

export interface TeacherProfileInput {
  name?: string;
  employeeId?: string;
  designation?: string;
  subject?: string;
  district?: string;
  block?: string;
  school?: string;
  schoolCode?: string;
  serviceYears?: number;
}

export interface TeacherListResponse {
  teachers: Teacher[];
  total: number;
  page: number;
  totalPages: number;
}

export interface TransferInput {
  targetTeacherId: number;
  message?: string;
}

export type TransferRequestStatus = typeof TransferRequestStatus[keyof typeof TransferRequestStatus];


export const TransferRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export interface TransferRequest {
  id: number;
  requesterId: number;
  targetId: number;
  requester?: Teacher;
  target?: Teacher;
  status: TransferRequestStatus;
  /** @nullable */
  message?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TransferStats {
  totalSent: number;
  totalReceived: number;
  pendingSent: number;
  pendingReceived: number;
  accepted: number;
  rejected: number;
  totalTeachersRegistered: number;
  recentActivity: TransferRequest[];
}

export type ListTeachersParams = {
district?: string;
subject?: string;
designation?: string;
block?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListTransfersParams = {
status?: ListTransfersStatus;
type?: ListTransfersType;
};

export type ListTransfersStatus = typeof ListTransfersStatus[keyof typeof ListTransfersStatus];


export const ListTransfersStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export type ListTransfersType = typeof ListTransfersType[keyof typeof ListTransfersType];


export const ListTransfersType = {
  sent: 'sent',
  received: 'received',
} as const;