# lib/api-zod/src/generated/types/mobileOtpRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bihar Teacher Mutual Transfer API
 * OpenAPI spec version: 0.1.0
 */

export interface MobileOtpRequest {
  /** 10-digit mobile number */
  phone: string;
}