# lib/api-zod/src/generated/types/mobileOtpVerify.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bihar Teacher Mutual Transfer API
 * OpenAPI spec version: 0.1.0
 */

export interface MobileOtpVerify {
  phone: string;
  otp: string;
}