# lib/api-zod/src/generated/types/teacher.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bihar Teacher Mutual Transfer API
 * OpenAPI spec version: 0.1.0
 */

export interface Teacher {
  id: number;
  name: string;
  employeeId: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** e.g. Primary Teacher, Middle Teacher, High School Teacher */
  designation: string;
  /** @nullable */
  subject?: string | null;
  district: string;
  block: string;
  school: string;
  /** @nullable */
  schoolCode?: string | null;
  /** @nullable */
  serviceYears?: number | null;
  isProfileComplete: boolean;
  createdAt: Date;
}